pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.3",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pikepdf>=9.10.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pypdf2>=3.0.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "httpx>=0.28.1",
]
